[CHORE] Data Module내의 파일들 네이밍 및 패키징 수정 (#50)
