+++ app/src/main/java/com/lgtm/android/di/RetrofitServiceModule.kt
@@ -1,6 +1,6 @@
 package com.lgtm.android.di
 
-import com.lgtm.android.data.remote.service.IntroService
+import com.lgtm.android.data.service.IntroService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
