+++ data/src/main/java/com/lgtm/android/data/model/response/IntroResponse.kt
@@ -1,10 +1,10 @@
-package com.lgtm.android.data.remote.model.response
+package com.lgtm.android.data.model.response
 
 import com.google.gson.annotations.SerializedName
 import com.lgtm.domain.constants.VERSION_UNKNOWN
 import com.lgtm.domain.entity.response.IntroVO
 
-data class IntroDTO(
+data class IntroResponse(
     @SerializedName("minVersion")
     val minVersion: Int?,
     @SerializedName("latestVersion")
